Rename to base_multi_image.
