+++ base_multi_image/__openerp__.py
