+++ base_multi_image/hooks.py
