+++ base_multi_image/i18n/sv.po
