+++ base_multi_image/security/ir.model.access.csv
@@ -1,0 +1,2 @@
+"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
+"access_images","Manage multi images","model_base_multi_image_image","base.group_user",1,1,1,1
