+++ base_multi_image/static/src/css/multi_image_base.sass
