+++ base_multi_image/views/assets.xml
@@ -8,7 +8,7 @@
 <template id="assets_frontend" inherit_id="website.assets_frontend">
     <xpath expr=".">
         <link rel="stylesheet"
-              href="/multi_image_base/static/src/css/multi_image_base.css"/>
+              href="/base_multi_image/static/src/css/base_multi_image.css"/>
     </xpath>
 </template>
