+++ base_multi_image/views/image_view.xml
@@ -4,7 +4,7 @@
 
 <record id="image_form_view" model="ir.ui.view">
     <field name="name">Multi image form</field>
-    <field name="model">multi_image_base.image</field>
+    <field name="model">base_multi_image.image</field>
     <field name="arch" type="xml">
         <form string="Image">
             <group>
@@ -65,7 +65,7 @@
 
 <record id="image_tree_view" model="ir.ui.view">
     <field name="name">Multi image tree</field>
-    <field name="model">multi_image_base.image</field>
+    <field name="model">base_multi_image.image</field>
     <field name="arch" type="xml">
         <tree string="Images">
             <field name="name"/>
@@ -78,7 +78,7 @@
 
 <record id="image_kanban_view" model="ir.ui.view">
     <field name="name">Product multi image kanban</field>
-    <field name="model">multi_image_base.image</field>
+    <field name="model">base_multi_image.image</field>
     <field name="arch" type="xml">
         <kanban string="Product Images">
             <field name="name"/>
@@ -93,7 +93,7 @@
                             <a type="open">
                                 <img
                                     t-att-src="kanban_image(
-                                        'multi_image_base.image',
+                                        'base_multi_image.image',
                                         'image_small',
                                         record.id.value)"
                                     class="oe_kanban_image"/>
@@ -126,7 +126,7 @@
 
 <record id="image_action" model="ir.actions.act_window">
     <field name="name">Multi images</field>
-    <field name="res_model">multi_image_base.image</field>
+    <field name="res_model">base_multi_image.image</field>
     <field name="view_mode">kanban,tree,form</field>
 </record>
