--- multi_image_base/security/ir.model.access.csv
@@ -1,2 +1,0 @@
-"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
-"access_images","Manage multi images","model_multi_image_base_image","base.group_user",1,1,1,1
